requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Utilities
numpy>=1.24.0
PyYAML>=6.0
dataclasses-json>=0.6.0

# Optional enhancements
pyttsx3>=2.90              # Text-to-speech (for voice output feature)

